[MB-14392] approveShipmentDiversion refactor and now includes ID
